@@ -23,6 +23,7 @@
     getInfo() {
       return {
         id: "0832rxfs",
+        // eslint-disable-next-line extension/should-translate
         name: "rxFS",
         color1: "#2bdab7",
         blocks: [
@@ -149,6 +150,7 @@
           {
             blockIconURI: file,
             opcode: "search",
+            hideFromPalette: true,
             blockType: Scratch.BlockType.REPORTER,
             text: Scratch.translate("search [STR]"),
             arguments: {
@@ -200,8 +202,11 @@
 
     del({ STR }) {
       str = btoa(unescape(encodeURIComponent(STR)));
-      rxFSfi[rxFSsy.indexOf(str) + 1 - 1] = undefined;
-      rxFSsy[rxFSsy.indexOf(str) + 1 - 1] = undefined;
+      const index = rxFSsy.indexOf(str);
+      if (index !== -1) {
+        rxFSfi.splice(index, 1);
+        rxFSsy.splice(index, 1);
+      }
     }
 
     file({ STR, STR2 }) {
