@@ -1,6 +1,8 @@
 // Name: Asset Manager
 // ID: lmsAssets
 // Description: Add, remove, and get data from various types of assets.
+// By: LilyMakesThings <https://scratch.mit.edu/users/LilyMakesThings/>
+// By: Mio <https://scratch.mit.edu/users/0znzw/>
 // License: MIT AND LGPL-3.0
 
 // TheShovel is so epic and cool and awesome
@@ -24,6 +26,10 @@
 
   class Assets {
     getInfo() {
+      const dataURIOption = Scratch.translate({
+        default: "dataURI",
+        description: "Menu option called dataURI",
+      });
       return {
         id: "lmsAssets",
         color1: "#5779ca",
@@ -174,9 +180,23 @@
             text: Scratch.translate("all sounds"),
           },
           {
+            // Legacy block
+            hideFromPalette: true,
             opcode: "getSpriteName",
             blockType: Scratch.BlockType.REPORTER,
             text: Scratch.translate("sprite name"),
+          },
+          {
+            disableMonitor: true,
+            opcode: "getSpriteValue",
+            blockType: Scratch.BlockType.REPORTER,
+            text: Scratch.translate("sprite [EXPORT]"),
+            arguments: {
+              EXPORT: {
+                type: Scratch.ArgumentType.STRING,
+                menu: "sprite",
+              },
+            },
           },
           "---",
           {
@@ -277,9 +297,23 @@
             },
           },
           {
+            // Legacy block
+            hideFromPalette: true,
             opcode: "getProjectJSON",
             blockType: Scratch.BlockType.REPORTER,
             text: Scratch.translate("project JSON"),
+          },
+          {
+            disableMonitor: true,
+            opcode: "getProjectValue",
+            blockType: Scratch.BlockType.REPORTER,
+            text: Scratch.translate("project [EXPORT]"),
+            arguments: {
+              EXPORT: {
+                type: Scratch.ArgumentType.STRING,
+                menu: "project",
+              },
+            },
           },
           "---",
           {
@@ -313,7 +347,7 @@
                 value: "index",
               },
               {
-                text: Scratch.translate("dataURI"),
+                text: dataURIOption,
                 value: "dataURI",
               },
               {
@@ -327,6 +361,32 @@
               {
                 text: Scratch.translate("asset ID"),
                 value: "asset ID",
+              },
+            ],
+          },
+          project: {
+            acceptReporters: false,
+            items: [
+              {
+                text: Scratch.translate("JSON"),
+                value: "JSON",
+              },
+              {
+                text: dataURIOption,
+                value: "dataURI",
+              },
+            ],
+          },
+          sprite: {
+            acceptReporters: false,
+            items: [
+              {
+                text: Scratch.translate("name"),
+                value: "name",
+              },
+              {
+                text: dataURIOption,
+                value: "dataURI",
               },
             ],
           },
@@ -362,8 +422,14 @@
       const assetType = this._typeIsBitmap(blob.type)
         ? runtime.storage.AssetType.ImageBitmap
         : runtime.storage.AssetType.ImageVector;
+
+      // Bitmap data format is not actually enforced, but setting it to something that isn't in scratch-parser's
+      // known format list will throw an error when someone tries to load the project.
+      // (https://github.com/scratchfoundation/scratch-parser/blob/665f05d739a202d565a4af70a201909393d456b2/lib/sb3_definitions.json#L51)
       const dataType =
-        blob.type === "image/svg+xml" ? "svg" : blob.type.split("/")[1];
+        blob.type === "image/svg+xml"
+          ? runtime.storage.DataFormat.SVG
+          : runtime.storage.DataFormat.PNG;
 
       const arrayBuffer = await new Promise((resolve, reject) => {
         const fr = new FileReader();
@@ -519,6 +585,29 @@
       return util.target.sprite.name ?? "";
     }
 
+    getSpriteValue(args, util) {
+      const option = Cast.toString(args.EXPORT);
+      if (option === "name") {
+        return util.target.sprite.name ?? "";
+      } else if (option === "dataURI") {
+        try {
+          return new Promise((resolve) => {
+            Scratch.vm.exportSprite(util.target.id).then((blob) => {
+              const fr = new FileReader();
+              fr.onload = () => resolve(fr.result);
+              fr.onabort = () => {
+                throw new Error("Read aborted");
+              };
+              fr.readAsDataURL(blob);
+            });
+          });
+        } catch (e) {
+          console.error("Failed to export the sprite", e);
+          return "";
+        }
+      }
+    }
+
     reorderCostume(args, util) {
       const target = util.target;
       const index1 = Cast.toNumber(args.INDEX1) - 1;
@@ -626,6 +715,29 @@
 
     getProjectJSON() {
       return Scratch.vm.toJSON();
+    }
+
+    getProjectValue(args) {
+      const option = Cast.toString(args.EXPORT);
+      if (option === "JSON") {
+        return Scratch.vm.toJSON();
+      } else if (option === "dataURI") {
+        try {
+          return new Promise((resolve) => {
+            vm.saveProjectSb3().then((blob) => {
+              const fr = new FileReader();
+              fr.onload = () => resolve(fr.result);
+              fr.onabort = () => {
+                throw new Error("Read aborted");
+              };
+              fr.readAsDataURL(blob);
+            });
+          });
+        } catch (e) {
+          console.error("Failed to export the project", e);
+          return "";
+        }
+      }
     }
 
     async loadExtension(args) {
