@@ -2,6 +2,8 @@
 // ID: lmsVideo
 // Description: Play videos from URLs.
 // By: LilyMakesThings <https://scratch.mit.edu/users/LilyMakesThings/>
+// By: SharkPool
+// By: Fath11 <https://scratch.mit.edu/users/fath11/>
 // License: MIT AND LGPL-3.0
 
 // Attribution is not required, but greatly appreciated.
@@ -14,6 +16,21 @@
   const renderer = vm.renderer;
   const Cast = Scratch.Cast;
 
+  // In some versions of Chrome, it seems that trying to render a <video> returns pure black
+  // if it's not in the DOM in a place the browser thinks is visible. That means we can't
+  // use display: none.
+  // See https://github.com/TurboWarp/scratch-render/issues/12
+  const elementContainer = document.createElement("div");
+  elementContainer.className = "tw-extensions-lily-videos-container";
+  elementContainer.style.pointerEvents = "none";
+  elementContainer.style.position = "absolute";
+  elementContainer.style.opacity = "0";
+  elementContainer.style.width = "0";
+  elementContainer.style.height = "0";
+  elementContainer.style.visibility = "hidden";
+  elementContainer.ariaHidden = "true";
+  document.body.appendChild(elementContainer);
+
   const BitmapSkin = runtime.renderer.exports.BitmapSkin;
   class VideoSkin extends BitmapSkin {
     constructor(id, renderer, videoName, videoSrc) {
@@ -24,6 +41,13 @@
 
       /** @type {string} */
       this.videoSrc = videoSrc;
+
+      /**
+       * Base volume as set by the scripts in the project, from 0 to 1.
+       * Does not account for eg. the project being muted.
+       * @type {number}
+       */
+      this.videoVolume = 1;
 
       this.videoError = false;
 
@@ -48,6 +72,10 @@
       };
       this.videoElement.src = videoSrc;
       this.videoElement.currentTime = 0;
+
+      // <video> must be in the DOM for it to render (see comments above)
+      elementContainer.appendChild(this.videoElement);
+      this.videoElement.tabIndex = -1;
 
       this.videoDirty = true;
 
@@ -88,6 +116,12 @@
       this.emitWasAltered();
     }
 
+    updateVolume() {
+      const projectVolume = runtime.audioEngine.inputNode.gain.value;
+      const trueVolume = this.videoVolume * projectVolume;
+      this.videoElement.volume = trueVolume;
+    }
+
     get size() {
       if (this.videoDirty) {
         this.reuploadVideo();
@@ -105,6 +139,7 @@
     dispose() {
       super.dispose();
       this.videoElement.pause();
+      this.videoElement.remove();
     }
   }
 
@@ -118,7 +153,28 @@
 
       runtime.on("BEFORE_EXECUTE", () => {
         for (const skin of renderer._allSkins) {
-          if (skin instanceof VideoSkin && !skin.videoElement.paused) {
+          if (skin instanceof VideoSkin) {
+            skin.updateVolume();
+            if (!skin.videoElement.paused) {
+              skin.markVideoDirty();
+            }
+          }
+        }
+      });
+
+      runtime.on("RUNTIME_PAUSED", () => {
+        for (const skin of renderer._allSkins) {
+          if (skin instanceof VideoSkin) {
+            skin.videoElement.pause();
+            skin.markVideoDirty();
+          }
+        }
+      });
+
+      runtime.on("RUNTIME_UNPAUSED", () => {
+        for (const skin of renderer._allSkins) {
+          if (skin instanceof VideoSkin) {
+            skin.videoElement.play();
             skin.markVideoDirty();
           }
         }
@@ -225,6 +281,23 @@
             },
           },
           {
+            opcode: "startVideoAndWait",
+            blockType: Scratch.BlockType.COMMAND,
+            text: Scratch.translate(
+              "start video [NAME] at [DURATION] seconds and wait until done"
+            ),
+            arguments: {
+              NAME: {
+                type: Scratch.ArgumentType.STRING,
+                defaultValue: "my video",
+              },
+              DURATION: {
+                type: Scratch.ArgumentType.NUMBER,
+                defaultValue: 0,
+              },
+            },
+          },
+          {
             opcode: "getAttribute",
             blockType: Scratch.BlockType.REPORTER,
             text: Scratch.translate("[ATTRIBUTE] of video [NAME]"),
@@ -233,6 +306,19 @@
                 type: Scratch.ArgumentType.STRING,
                 menu: "attribute",
               },
+              NAME: {
+                type: Scratch.ArgumentType.STRING,
+                defaultValue: "my video",
+              },
+            },
+          },
+          {
+            opcode: "getFrame",
+            blockType: Scratch.BlockType.REPORTER,
+            text: Scratch.translate(
+              "screenshot of video [NAME] at current time"
+            ),
+            arguments: {
               NAME: {
                 type: Scratch.ArgumentType.STRING,
                 defaultValue: "my video",
@@ -259,6 +345,21 @@
               NAME: {
                 type: Scratch.ArgumentType.STRING,
                 defaultValue: "my video",
+              },
+            },
+          },
+          {
+            opcode: "toggleLooping",
+            blockType: Scratch.BlockType.COMMAND,
+            text: Scratch.translate("set video [NAME] to [LOOP]"),
+            arguments: {
+              NAME: {
+                type: Scratch.ArgumentType.STRING,
+                defaultValue: "my video",
+              },
+              LOOP: {
+                type: Scratch.ArgumentType.STRING,
+                menu: "playbackType",
               },
             },
           },
@@ -293,6 +394,23 @@
               },
             },
           },
+          {
+            opcode: "setPlaybackRate",
+            blockType: Scratch.BlockType.COMMAND,
+            text: Scratch.translate(
+              "set playback rate of video [NAME] to [RATE]"
+            ),
+            arguments: {
+              NAME: {
+                type: Scratch.ArgumentType.STRING,
+                defaultValue: "my video",
+              },
+              RATE: {
+                type: Scratch.ArgumentType.NUMBER,
+                defaultValue: "2",
+              },
+            },
+          },
         ],
         menus: {
           targets: {
@@ -309,6 +427,10 @@
               {
                 text: Scratch.translate("paused"),
                 value: "paused",
+              },
+              {
+                text: Scratch.translate("looping"),
+                value: "looping",
               },
             ],
           },
@@ -334,6 +456,23 @@
               {
                 text: Scratch.translate("height"),
                 value: "height",
+              },
+              {
+                text: Scratch.translate("playback rate"),
+                value: "playback rate",
+              },
+            ],
+          },
+          playbackType: {
+            acceptReporters: false,
+            items: [
+              {
+                text: Scratch.translate("loop"),
+                value: "loop",
+              },
+              {
+                text: Scratch.translate("not loop"),
+                value: "not loop",
               },
             ],
           },
@@ -445,6 +584,25 @@
       videoSkin.markVideoDirty();
     }
 
+    startVideoAndWait(args, util) {
+      const videoName = Cast.toString(args.NAME);
+      const duration = Cast.toNumber(args.DURATION);
+      const videoSkin = this.videos[videoName];
+      if (!videoSkin) return;
+
+      if (!util.stackFrame.hasPlayed) {
+        videoSkin.videoElement.play();
+        videoSkin.videoElement.currentTime = duration;
+        videoSkin.markVideoDirty();
+
+        util.stackFrame.hasPlayed = true;
+      }
+
+      if (!videoSkin.videoElement.ended) {
+        util.yield();
+      }
+    }
+
     getAttribute(args) {
       const videoName = Cast.toString(args.NAME);
       const videoSkin = this.videos[videoName];
@@ -456,14 +614,39 @@
         case "duration":
           return videoSkin.videoElement.duration;
         case "volume":
-          return videoSkin.videoElement.volume * 100;
+          return videoSkin.videoVolume * 100;
         case "width":
           return videoSkin.size[0];
         case "height":
           return videoSkin.size[1];
+        case "playback rate":
+          return videoSkin.videoElement.playbackRate;
         default:
           return 0;
       }
+    }
+
+    getFrame(args) {
+      const videoName = Cast.toString(args.NAME);
+      const videoSkin = this.videos[videoName];
+      if (!videoSkin) return "";
+
+      const canvas = document.createElement("canvas");
+      const ctx = canvas.getContext("2d");
+      if (!ctx) {
+        console.warn("2D rendering context not available");
+        return "";
+      }
+
+      const videoElement = videoSkin.videoElement;
+      if (videoElement.videoWidth === 0 || videoElement.videoHeight === 0) {
+        return "";
+      }
+
+      canvas.width = videoElement.videoWidth;
+      canvas.height = videoElement.videoHeight;
+      ctx.drawImage(videoElement, 0, 0);
+      return canvas.toDataURL();
     }
 
     pause(args) {
@@ -484,23 +667,54 @@
       videoSkin.markVideoDirty();
     }
 
+    toggleLooping(args) {
+      const videoName = Cast.toString(args.NAME);
+      const videoSkin = this.videos[videoName];
+      if (!videoSkin) return;
+
+      videoSkin.videoElement.loop = args.LOOP == "loop" ? true : false;
+    }
+
     getState(args) {
       const videoName = Cast.toString(args.NAME);
       const videoSkin = this.videos[videoName];
       if (!videoSkin) return args.STATE === "paused";
 
-      return args.STATE == "playing"
-        ? !videoSkin.videoElement.paused
-        : videoSkin.videoElement.paused;
+      switch (args.STATE) {
+        case "playing":
+          return !videoSkin.videoElement.paused;
+        case "paused":
+          return videoSkin.videoElement.paused;
+        case "looping":
+          return videoSkin.videoElement.loop;
+        default:
+          return false;
+      }
     }
 
     setVolume(args) {
       const videoName = Cast.toString(args.NAME);
-      const value = Cast.toNumber(args.VALUE);
       const videoSkin = this.videos[videoName];
       if (!videoSkin) return;
 
-      videoSkin.videoElement.volume = value / 100;
+      const value = Cast.toNumber(args.VALUE);
+      videoSkin.videoVolume = Math.min(1, Math.max(0, value / 100));
+      videoSkin.updateVolume();
+    }
+
+    setPlaybackRate(args) {
+      const videoName = Cast.toString(args.NAME);
+      const videoSkin = this.videos[videoName];
+      if (!videoSkin) return;
+
+      try {
+        const value = Cast.toNumber(args.RATE);
+        // Supposedly negative values will work in Safari but people probably shouldn't rely
+        // on that since others don't.
+        videoSkin.videoElement.playbackRate = Math.max(0, value);
+      } catch (e) {
+        console.warn(e);
+      }
     }
 
     /** @returns {VM.Target|undefined} */
