@@ -148,6 +148,24 @@
               },
             },
             extensions: ["colours_motion"],
+            hideFromPalette: true,
+          },
+          {
+            filter: [Scratch.TargetType.SPRITE],
+            opcode: "directionto2",
+            blockType: Scratch.BlockType.REPORTER,
+            text: Scratch.translate("direction to x: [X] y: [Y]"),
+            arguments: {
+              X: {
+                type: Scratch.ArgumentType.NUMBER,
+                defaultValue: "0",
+              },
+              Y: {
+                type: Scratch.ArgumentType.NUMBER,
+                defaultValue: "0",
+              },
+            },
+            extensions: ["colours_motion"],
           },
           {
             filter: [Scratch.TargetType.SPRITE],
@@ -257,21 +275,12 @@
       util.target.setXY(util.target.x + x, util.target.y + y);
     }
 
-    // LORAX APPROVED
     pointto(args, util) {
       const x = Scratch.Cast.toNumber(args.X);
       const y = Scratch.Cast.toNumber(args.Y);
-      if (util.target.y > y) {
-        util.target.setDirection(
-          (180 / Math.PI) *
-            Math.atan((x - util.target.x) / (y - util.target.y)) +
-            180
-        );
-      } else {
-        util.target.setDirection(
-          (180 / Math.PI) * Math.atan((x - util.target.x) / (y - util.target.y))
-        );
-      }
+      util.target.setDirection(
+        (180 / Math.PI) * Math.atan2(x - util.target.x, y - util.target.y)
+      );
     }
 
     rotationStyle(args, util) {
@@ -287,6 +296,7 @@
     }
 
     directionto(args, util) {
+      // Old version, returns values from -90 to 270
       const x = Scratch.Cast.toNumber(args.X);
       const y = Scratch.Cast.toNumber(args.Y);
       if (util.target.y > y) {
@@ -300,6 +310,13 @@
           (180 / Math.PI) * Math.atan((x - util.target.x) / (y - util.target.y))
         );
       }
+    }
+
+    directionto2(args, util) {
+      // New version, returns values from -180 to 180, like Scratch direction reporter.
+      const x = Scratch.Cast.toNumber(args.X);
+      const y = Scratch.Cast.toNumber(args.Y);
+      return (180 / Math.PI) * Math.atan2(x - util.target.x, y - util.target.y);
     }
 
     distanceto(args, util) {
