@@ -34,7 +34,7 @@
     let ws;
     try {
       // Permission is checked earlier.
-      // eslint-disable-next-line no-restricted-syntax
+      // eslint-disable-next-line extension/check-can-fetch
       ws = new WebSocket(uri);
     } catch (e) {
       return {
