@@ -105,22 +105,22 @@
     var t = v * (1 - (1 - f) * s);
     switch (i % 6) {
       case 0:
-        (r = v), (g = t), (b = p);
+        ((r = v), (g = t), (b = p));
         break;
       case 1:
-        (r = q), (g = v), (b = p);
+        ((r = q), (g = v), (b = p));
         break;
       case 2:
-        (r = p), (g = v), (b = t);
+        ((r = p), (g = v), (b = t));
         break;
       case 3:
-        (r = p), (g = q), (b = v);
+        ((r = p), (g = q), (b = v));
         break;
       case 4:
-        (r = t), (g = p), (b = v);
+        ((r = t), (g = p), (b = v));
         break;
       case 5:
-        (r = v), (g = p), (b = q);
+        ((r = v), (g = p), (b = q));
         break;
     }
     return [(r * 255) | 0, (g * 255) | 0, (b * 255) | 0];
@@ -604,6 +604,13 @@
 
       return this._texture;
     }
+
+    isMetricsReady() {
+      if (this._needsReflow()) {
+        this._reflowText();
+      }
+      return true;
+    }
   }
 
   /**
@@ -836,7 +843,7 @@
           {
             opcode: "setOutlineWidth",
             blockType: Scratch.BlockType.COMMAND,
-            text: "set outline width to [WIDTH]",
+            text: Scratch.translate("set outline width to [WIDTH]"),
             hideFromPalette: compatibilityMode,
             arguments: {
               WIDTH: {
@@ -849,7 +856,7 @@
           {
             opcode: "setOutlineColor",
             blockType: Scratch.BlockType.COMMAND,
-            text: "set outline color to [COLOR]",
+            text: Scratch.translate("set outline color to [COLOR]"),
             hideFromPalette: compatibilityMode,
             arguments: {
               COLOR: {
@@ -1021,7 +1028,7 @@
           {
             opcode: "setShakeIntensity",
             blockType: Scratch.BlockType.COMMAND,
-            text: "set shake intensity to [NUM]%",
+            text: Scratch.translate("set shake intensity to [NUM]%"),
             hideFromPalette: compatibilityMode,
             arguments: {
               NUM: {
